requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Profile & Resume Parsing
pdfminer.six>=20221105
python-docx>=1.0.0

# NLP & Analysis
spacy>=3.7.0
scikit-learn>=1.4.0
rapidfuzz>=3.6.0

# Resume Generation
jinja2>=3.1.0
weasyprint>=61.0

# Browser Automation
playwright>=1.41.0

# Scheduling
apscheduler>=3.10.0

# Credentials
keyring>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Notifications
python-telegram-bot>=21.0
